requirements.txt
pytest==8.1.0
selenium==4.18.1
requests==2.31.0
faker==23.3.0
allure-pytest==2.13.2
pytest-xdist